+++ RELEASES.md
@@ -9,7 +9,7 @@
 
 ### Changed
 
-* [msr] Updated for the Intel Software Developer's Manual, December 2023 release
+* [msr] Updated for the Intel Software Developer's Manual, March 2024 release
 * [rapl-configure] Refactor help output and man page documentation on zones and constraints
 
 ### Removed
@@ -19,6 +19,7 @@
 ### Fixed
 
 * [msr] Incorrect maximum power limit for Sapphire Rapids PSYS zone
+* [msr] Sapphire Rapids and Emerald Rapids use 61 uJ DRAM energy units (per Intel SDM, March 2024 release)
 
 
 ## [v0.9.1] - 2023-11-19